requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
yfinance
panda
numpy